# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mi Negocio Pro API",
        "description": "API para gestión completa de negocio: usuarios, productos, ventas, compras y más",
        "contact": {
            "email": "support@minegocio.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server Local"
        }
    ],
    "paths": {
...
        "/api/purchases": {
            "get": {
                "tags": ["Purchases"],
                "summary": "Listar compras",
                "description": "Obtiene la lista de compras con filtros opcionales",
                "operationId": "getPurchases",
                "parameters": [
                    {
                        "name": "date_from",
                        "in": "query",
                        "description": "Fecha desde (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "description": "Fecha hasta (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "received_by",
                        "in": "query",
                        "description": "Filtrar por ID de usuario que recibió",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "supplier_name",
                        "in": "query",
                        "description": "Filtrar por nombre de proveedor",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Número de elementos por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15,
                            "maximum": 100,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de compras",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "total": {
                                                    "type": "integer"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "last_page": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": ["Purchases"],
                "summary": "Crear nueva compra",
                "description": "Crea una nueva compra/entrada de inventario",
                "operationId": "createPurchase",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": ["received_by"],
                                "properties": {
                                    "supplier_name": {
                                        "description": "Nombre del proveedor",
                                        "type": "string",
                                        "example": "Distribuidora ABC"
                                    },
                                    "notes": {
                                        "description": "Notas adicionales",
                                        "type": "string",
                                        "example": "Compra de productos varios"
                                    },
                                    "purchase_date": {
                                        "description": "Fecha de la compra",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-08-24"
                                    },
                                    "received_by": {
                                        "description": "ID del usuario que recibió",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Compra creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Compra creada exitosamente"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Datos inválidos"
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "403": {
                        "description": "Prohibido"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/purchases/stats": {
            "get": {
                "tags": ["Purchases"],
                "summary": "Estadísticas de compras",
                "description": "Obtiene estadísticas de compras con filtros opcionales",
                "operationId": "getPurchasesStats",
                "parameters": [
                    {
                        "name": "date_from",
                        "in": "query",
                        "description": "Fecha desde (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "description": "Fecha hasta (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "received_by",
                        "in": "query",
                        "description": "Filtrar por ID de usuario",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "supplier_name",
                        "in": "query",
                        "description": "Filtrar por nombre de proveedor",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Estadísticas de compras",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "total_purchases": {
                                                    "type": "integer"
                                                },
                                                "total_amount": {
                                                    "type": "number",
                                                    "format": "decimal"
                                                },
                                                "average_amount": {
                                                    "type": "number",
                                                    "format": "decimal"
                                                },
                                                "monthly_purchases": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "top_suppliers": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}